NCI-Agency/anet#561: Make first form fieldset title the form title
This moves the title of the first fieldset of the forms one level up to
become title of the form (on the same line as the top submit button).
Make this form title sticky as we do not want the title and the submit
button to dissapear when we scroll.